requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
anthropic==0.8.1
openai>=1.6.1
flask-cors==4.0.0
pyJWT==2.8.0